convert verifyMaci to a hardhat task
